allow more filters for checking whether a transaction should be processed or not.
also updated the readme
